Default to SNMP v2c communication
The poll device `hcounters` attribute is really what drives the SNMP
version selection in Zino.  Unfortunately, the default is SNMP v1, and
this has worked because PySNMP silently allowed v2c requests like
BULK-GET to be sent over sessions that were otherwise v1.

Net-SNMP does not allow the sending of v2c PDUs over v1 sessions, so
we need to be more strict about this.  It also makes no sense to default
to v1 in 2025, so this sets `hcounters`' default value to True and
fixes a test that implicitly assumed v1 exceptions to be raised.